python/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-stubs==5.0.4
django-stubs-ext==5.0.4
djangorestframework==3.15.2
djangorestframework-stubs==3.15.0
idna==3.8
mypy==1.11.2
mypy-extensions==1.0.0
requests==2.32.3
sqlparse==0.5.1
types-PyYAML==6.0.12.20240808
types-requests==2.32.0.20240907
typing_extensions==4.12.2
urllib3==2.2.2
